cardano-cli | submit-tx:  new command & some refactoring
