+++ cardano-node/src/Cardano/Node/Parsers.hs
@@ -5,6 +5,7 @@ module Cardano.Node.Parsers
   ) where
 
 
+import           Prelude (String, error, id)
 import           Cardano.Prelude hiding (option)
 
 import           Cardano.Common.Protocol (Protocol(..))
@@ -23,6 +24,28 @@ parseCoreNodeId =
          <> metavar "CORE-NODE-ID"
          <> help "The ID of the core node to which this client is connected."
     )
+
+parseNodeId :: String -> Parser NodeId
+parseNodeId desc =
+    option (fmap CoreId auto) (
+            long "node-id"
+         <> short 'n'
+         <> metavar "NODE-ID"
+         <> help desc
+    )
+
+parseTopologyInfo :: String -> Parser TopologyInfo
+parseTopologyInfo desc = TopologyInfo <$> parseNodeId desc <*> parseTopologyFile
+
+parseTopologyFile :: Parser FilePath
+parseTopologyFile =
+    strOption (
+            long "topology"
+         <> short 't'
+         <> metavar "FILEPATH"
+         <> help "The path to a file describing the topology."
+    )
+
 -- | The parser for the logging specific arguments.
 loggingParser :: Parser LoggingCLIArguments
 loggingParser = LoggingCLIArguments
@@ -33,22 +56,46 @@ loggingParser = LoggingCLIArguments
        <> completer (bashCompleter "file")
         )
 
-parseProtocol :: Parser Protocol
-parseProtocol = asum [
-      flag' BFT $ mconcat [
-          long "bft"
-        , help "Use the BFT consensus algorithm"
+parseLovelace :: String -> String -> Parser Lovelace
+parseLovelace optname desc =
+  either (error . show) id . mkLovelace
+  <$> parseIntegral optname desc
+
+parseLovelacePortion :: String -> String -> Parser LovelacePortion
+parseLovelacePortion optname desc =
+  either (error . show) id . mkLovelacePortion
+  <$> parseIntegral optname desc
+
+parseFakeAvvmOptions :: Parser Genesis.FakeAvvmOptions
+parseFakeAvvmOptions =
+  Genesis.FakeAvvmOptions
+  <$> parseIntegral        "avvm-entry-count"         "Number of AVVM addresses."
+  <*> parseLovelace        "avvm-entry-balance"       "AVVM address."
+
+parseK :: Parser BlockCount
+parseK =
+  BlockCount
+  <$> parseIntegral        "k"                        "The security parameter of the Ouroboros protocol."
+
+parseProtocolMagicId :: String -> Parser ProtocolMagicId
+parseProtocolMagicId arg =
+  ProtocolMagicId
+  <$> parseIntegral        arg                        "The magic number unique to any instance of Cardano."
+
+parseProtocolMagic :: Parser ProtocolMagic
+parseProtocolMagic =
+  flip AProtocolMagic RequiresMagic . flip Annotated ()
+  <$> parseProtocolMagicId "protocol-magic"
+
+parseNetworkMagic :: Parser NetworkMagic
+parseNetworkMagic = asum
+    [ flag' NetworkMainOrStage $ mconcat [
+          long "main-or-staging"
+        , help ""
         ]
-    , flag' Praos $ mconcat [
-          long "praos"
-        , help "Use the Praos consensus algorithm"
-        ]
-    , flag' MockPBFT $ mconcat [
-          long "mock-pbft"
-        , help "Use the Permissive BFT consensus algorithm using a mock ledger"
-        ]
-    , flag' RealPBFT $ mconcat [
-          long "real-pbft"
-        , help "Use the Permissive BFT consensus algorithm using the real ledger"
-        ]
+    , option (fmap NetworkTestnet auto) (
+          long "testnet-magic"
+       <> metavar "MAGIC"
+       <> help "The testnet network magic, decibal"
+        )
     ]
